Add files via upload
Python script to check graphs with suturing sequences, requires the file "suturing_file_sequences.csv" and the maude binary to be in the same folder. 
Preliminary results, it is nice to have results but the graphs and sequences are not entirely correct yet, for now it is just a demonstration that our approach works and we can have results. The sequences and graphs can be easily corrected.